# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API doc of UniqueStudio Recruitment. For more API information, please see https://app.apifox.com/project/2985744",
        "title": "UniqueStudio Recruitment API",
        "contact": {
            "email": "wwbstar07@gmail.com"
        },
        "version": "0.1"
    },
    "host": "https://dev.back.recruitment2023.hustunique.com/",
    "paths": {
...
        "/recruitments": {
            "get": {
                "description": "get pending(the most recent) recruitment, member can only get the recruitment's detail(include application, interviews) after join in.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recruitment"
                ],
                "summary": "get pending recruitment.",
                "operationId": "get_pending_recruitment.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pkg.Recruitment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            },
            "post": {
                "description": "gcreate recruitment, only can be created by admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recruitment"
                ],
                "summary": "create recruitment.",
                "operationId": "create_recruitment.",
                "parameters": [
                    {
                        "description": "create recruitment opts",
                        "name": "pkg.CreateRecOpts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.CreateRecOpts"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pkg.Recruitment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
        "/recruitments/{rid}": {
            "get": {
                "description": "get recruitment, member can only get the recruitment's detail(include application, interviews) after join in.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recruitment"
                ],
                "summary": "get recruitment.",
                "operationId": "get_recruitment.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recruitment uid",
                        "name": "rid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pkg.Recruitment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            },
            "put": {
                "description": "update recruitment, only can be updated by admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recruitment"
                ],
                "summary": "update recruitment.",
                "operationId": "update_recruitment.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recruitment uid",
                        "name": "rid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update recruitment opts",
                        "name": "pkg.UpdateRecOpts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.UpdateRecOpts"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
        "/recruitments/{rid}/file/{group}/{type}": {
            "get": {
                "description": "upload recruitment file, such as written test.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recruitment"
                ],
                "summary": "download recruitment file, such as written test.",
                "operationId": "download_recruitment_file",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            },
            "put": {
                "description": "upload recruitment file, such as written test.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recruitment"
                ],
                "summary": "upload recruitment file, such as written test.",
                "operationId": "upload_recruitment_file",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
...
        "/recruitments/{rid}/stressTest": {
            "put": {
                "description": "set stress test start and end time.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recruitment"
                ],
                "summary": "set stress test start and end time.",
                "operationId": "set_stress_test_time",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "pkg.Application": {
            "type": "object",
            "properties": {
                "abandoned": {
                    "type": "boolean"
                },
                "answer": {
                    "type": "string"
                },
                "candidate_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "comments": {
                    "description": "onetomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg.Comment"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "grade": {
                    "description": "pkg.Grade",
                    "type": "string"
                },
                "group": {
                    "description": "pkg.Group",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pkg.Group"
                        }
                    ]
                },
                "institute": {
                    "type": "string"
                },
                "interview_allocations_group": {
                    "$ref": "#/definitions/pkg.Interview"
                },
                "interview_allocations_group_id": {
                    "type": "string"
                },
                "interview_allocations_team": {
                    "$ref": "#/definitions/pkg.Interview"
                },
                "interview_allocations_team_id": {
                    "type": "string"
                },
                "interview_selections": {
                    "description": "manytomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg.Interview"
                    }
                },
                "intro": {
                    "type": "string"
                },
                "is_quick": {
                    "type": "boolean"
                },
                "major": {
                    "type": "string"
                },
                "rank": {
                    "type": "string"
                },
                "recruitment_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "referrer": {
                    "type": "string"
                },
                "rejected": {
                    "type": "boolean"
                },
                "resume": {
                    "type": "string"
                },
                "step": {
                    "description": "pkg.Step",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pkg.Step"
                        }
                    ]
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_detail": {
                    "description": "get from sso",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pkg.UserDetail"
                        }
                    ]
                }
            }
        },
        "pkg.Comment": {
            "type": "object",
            "properties": {
                "application_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "evaluation": {
                    "$ref": "#/definitions/pkg.Evaluation"
                },
                "member_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "member_name": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "pkg.CreateRecOpts": {
            "type": "object",
            "required": [
                "beginning",
                "deadline",
                "end",
                "name"
            ],
            "properties": {
                "beginning": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "pkg.Evaluation": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "Good",
                "Normal",
                "Bad"
            ]
        },
        "pkg.Gender": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "Male",
                "Female",
                "Oth"
            ]
        },
        "pkg.Group": {
            "type": "string",
            "enum": [
                "web",
                "lab",
                "ai",
                "game",
                "mobile",
                "android",
                "ios",
                "design",
                "pm",
                "unique"
            ],
            "x-enum-comments": {
                "Unique": "for team interview"
            },
            "x-enum-varnames": [
                "Web",
                "Lab",
                "Ai",
                "Game",
                "Mobile",
                "Android",
                "Ios",
                "Design",
                "Pm",
                "Unique"
            ]
        },
        "pkg.Interview": {
            "type": "object",
            "properties": {
                "applications": {
                    "description": "manytomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg.Application"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "name": {
                    "$ref": "#/definitions/pkg.Group"
                },
                "period": {
                    "$ref": "#/definitions/pkg.Period"
                },
                "recruitment_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "pkg.Period": {
            "type": "string",
            "enum": [
                "morning",
                "afternoon",
                "evening"
            ],
            "x-enum-varnames": [
                "Morning",
                "Afternoon",
                "Evening"
            ]
        },
        "pkg.Recruitment": {
            "type": "object",
            "properties": {
                "applications": {
                    "description": "一个hr-\u003e简历 ;级联删除",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg.Application"
                    }
                },
                "beginning": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "group_details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "interviews": {
                    "description": "一个hr-\u003e面试 ;级联删除",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg.Interview"
                    }
                },
                "name": {
                    "type": "string"
                },
                "statistics": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "stress_test_end": {
                    "type": "string"
                },
                "stress_test_start": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "pkg.Step": {
            "type": "string",
            "enum": [
                "SignUp",
                "WrittenTest",
                "GroupTimeSelection",
                "GroupInterview",
                "OnlineGroupInterview",
                "StressTest",
                "TeamTimeSelection",
                "TeamInterview",
                "OnlineTeamInterview",
                "Pass"
            ],
            "x-enum-comments": {
                "GroupInterview": "组面",
                "GroupTimeSelection": "组面时间选择",
                "OnlineGroupInterview": "在线组面",
                "OnlineTeamInterview": "在线群面",
                "Pass": "通过",
                "SignUp": "报名",
                "StressTest": "熬测",
                "TeamInterview": "群面",
                "TeamTimeSelection": "面试时间选择",
                "WrittenTest": "笔试"
            },
            "x-enum-varnames": [
                "SignUp",
                "WrittenTest",
                "GroupTimeSelection",
                "GroupInterview",
                "OnlineGroupInterview",
                "StressTest",
                "TeamTimeSelection",
                "TeamInterview",
                "OnlineTeamInterview",
                "Pass"
            ]
        },
...
        "pkg.UpdateRecOpts": {
            "type": "object",
            "properties": {
                "beginning": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rid": {
                    "type": "string"
                }
            }
        },
        "pkg.UserDetail": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/pkg.Gender"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "join_time": {
                    "type": "string"
                },
                "lark_union_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
    },
...
}